Extract the script from MakersShowView.vue into a composable
